Rename order book reconstruction
Currently the order book reconstruction provides an aggregate
view of the best price level (best bid and offer, BBO). Rename
it in preparation of enhancing it to provide an aggregate view
of all price levels.
