+++ parity-book-perf-test/README.md
@@ -1,0 +1,20 @@
+Parity Order Book Performance Test
+==================================
+
+Parity Order Book Performance Test contains microbenchmarks for Parity Order
+Book.
+
+
+Usage
+-----
+
+Run Parity Order Book Performance Test with Java:
+
+    java -jar <executable>
+
+
+License
+-------
+
+Parity Order Book Performance Test is released under the Apache License,
+Version 2.0.
