+++ parity-book/README.md
@@ -1,0 +1,23 @@
+Parity Order Book
+=================
+
+Parity Order Book implements high-performance order book reconstruction on the
+JVM.
+
+
+Download
+--------
+
+Add a Maven dependency to Parity Order Book:
+
+    <dependency>
+      <groupId>com.paritytrading.parity</groupId>
+      <artifactId>parity-book</artifactId>
+      <version><!-- latest version --></version>
+    </dependency>
+
+
+License
+-------
+
+Parity Order Book is released under the Apache License, Version 2.0.
