+++ parity-book/src/main/java/com/paritytrading/parity/book/AskComparator.java
@@ -1,0 +1,14 @@
+package com.paritytrading.parity.book;
+
+import it.unimi.dsi.fastutil.longs.AbstractLongComparator;
+
+class AskComparator extends AbstractLongComparator {
+
+    public static final AskComparator INSTANCE = new AskComparator();
+
+    @Override
+    public int compare(long left, long right) {
+        return Long.signum(left - right);
+    }
+
+}
