+++ parity-book/src/main/java/com/paritytrading/parity/book/BidComparator.java
@@ -1,0 +1,14 @@
+package com.paritytrading.parity.book;
+
+import it.unimi.dsi.fastutil.longs.AbstractLongComparator;
+
+class BidComparator extends AbstractLongComparator {
+
+    public static final BidComparator INSTANCE = new BidComparator();
+
+    @Override
+    public int compare(long left, long right) {
+        return Long.signum(right - left);
+    }
+
+}
