+++ parity-book/src/main/java/com/paritytrading/parity/book/Side.java
@@ -1,0 +1,9 @@
+package com.paritytrading.parity.book;
+
+/**
+ * The order side.
+ */
+public enum Side {
+    BUY,
+    SELL
+}
