+++ parity-book/src/main/java/com/paritytrading/parity/book/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Parity Order Book implements high-performance order book reconstruction.
+ */
+package com.paritytrading.parity.book;
