+++ parity-ticker/src/main/java/com/paritytrading/parity/ticker/ListenCommand.java
@@ -6,8 +6,8 @@ import com.paritytrading.foundation.ASCII;
 import com.paritytrading.nassau.MessageListener;
 import com.paritytrading.nassau.util.MoldUDP64;
 import com.paritytrading.nassau.util.SoupBinTCP;
+import com.paritytrading.parity.book.Market;
 import com.paritytrading.parity.net.pmd.PMDParser;
-import com.paritytrading.parity.top.Market;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigException;
 import java.io.FileNotFoundException;
