+++ parity-ticker/src/main/java/com/paritytrading/parity/ticker/MarketDataProcessor.java
@@ -2,10 +2,10 @@ package com.paritytrading.parity.ticker;
 
 import static org.jvirtanen.util.Applications.*;
 
+import com.paritytrading.parity.book.Market;
+import com.paritytrading.parity.book.Side;
 import com.paritytrading.parity.net.pmd.PMD;
 import com.paritytrading.parity.net.pmd.PMDListener;
-import com.paritytrading.parity.top.Market;
-import com.paritytrading.parity.top.Side;
 
 class MarketDataProcessor implements PMDListener {
