+++ parity-ticker/src/main/java/com/paritytrading/parity/ticker/ReadCommand.java
@@ -4,8 +4,8 @@ import static org.jvirtanen.util.Applications.*;
 
 import com.paritytrading.foundation.ASCII;
 import com.paritytrading.nassau.util.BinaryFILE;
+import com.paritytrading.parity.book.Market;
 import com.paritytrading.parity.net.pmd.PMDParser;
-import com.paritytrading.parity.top.Market;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
