--- parity-top-perf-test/README.md
@@ -1,20 +1,0 @@
-Parity Top of Book Performance Test
-===================================
-
-Parity Top of Book Performance Test contains microbenchmarks for Parity Top of
-Book.
-
-
-Usage
------
-
-Run Parity Top of Book Performance Test with Java:
-
-    java -jar <executable>
-
-
-License
--------
-
-Parity Top of Book Performance Test is released under the Apache License,
-Version 2.0.
