--- parity-top/README.md
@@ -1,23 +1,0 @@
-Parity Top of Book
-==================
-
-Parity Top of Book implements high-performance order book reconstruction on the
-JVM.
-
-
-Download
---------
-
-Add a Maven dependency to Parity Top of Book:
-
-    <dependency>
-      <groupId>com.paritytrading.parity</groupId>
-      <artifactId>parity-top</artifactId>
-      <version><!-- latest version --></version>
-    </dependency>
-
-
-License
--------
-
-Parity Top of Book is released under the Apache License, Version 2.0.
