--- parity-top/pom.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>com.paritytrading.parity</groupId>
-    <artifactId>parity-parent</artifactId>
-    <version>0.5.1-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>parity-top</artifactId>
-
-  <name>Parity Top of Book</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>it.unimi.dsi</groupId>
-      <artifactId>fastutil</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jvirtanen.value</groupId>
-      <artifactId>value</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-</project>
