--- parity-top/src/main/java/com/paritytrading/parity/top/BidComparator.java
@@ -1,14 +1,0 @@
-package com.paritytrading.parity.top;
-
-import it.unimi.dsi.fastutil.longs.AbstractLongComparator;
-
-class BidComparator extends AbstractLongComparator {
-
-    public static final BidComparator INSTANCE = new BidComparator();
-
-    @Override
-    public int compare(long left, long right) {
-        return Long.signum(right - left);
-    }
-
-}
