--- parity-top/src/main/java/com/paritytrading/parity/top/Side.java
@@ -1,9 +1,0 @@
-package com.paritytrading.parity.top;
-
-/**
- * The order side.
- */
-public enum Side {
-    BUY,
-    SELL
-}
