--- parity-top/src/main/java/com/paritytrading/parity/top/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Parity Top of Book implements high-performance order book reconstruction.
- */
-package com.paritytrading.parity.top;
